pyproject.toml
[project]
name = "mcp-server-hive"
version = "0.1.0"
description = "MCP Hive Server for database operations"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv",
    "mcp",
    "openai",
    "pyhive",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true
